--- nixos/lib/make-options-doc/optionsJSONtoXML.nix
@@ -1,6 +1,0 @@
-{ file }:
-
-builtins.attrValues
-  (builtins.mapAttrs
-    (name: def: def // { inherit name; })
-    (builtins.fromJSON (builtins.readFile file)))
